Implement gameController function
gameController controls the flow of the game.
gameController creates the board and players. It
contains the switchPlayer function to change turns
and the playRound function to get one player to
play. If player plays legal move in an empty cell,
gameController will change turns. Otherwise, turns
will not change. gameController returns an object
containing the playRound function.